requirements.txt
streamlit
torch
torchvision
Pillow
python-docx
docx2pdf
